Add shard annotation and make associated changes for annotation driver
Added annotation "shard" allows to mark a property which data is placed
in a shard. A shard manager was passed into an entity manager. It helps
us to turn on a target shard and then come back the old one in base
entity persister. At the moment it works only for OneToMany annotation.